switch to `JSTraceable`
